% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core Functions.R, R/enrichment.R
\name{enrich_analysis}
\alias{enrich_analysis}
\title{GO and KEGG enrichment analysis}
\usage{
enrich_analysis(genes, OrgDb)

enrich_analysis(genes, OrgDb)

enrich_analysis(genes, OrgDb)
}
\arguments{
\item{genes}{Character vector of Entrez gene IDs.}

\item{OrgDb}{AnnotationDb object, e.g. org.Hs.eg.db.}
}
\value{
List with GO and KEGG results and ggplot objects.

List with GO and KEGG results and ggplot objects.
}
\description{
GO and KEGG enrichment analysis

GO/KEGG enrichment

GO and KEGG enrichment analysis
}
