% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core Functions.R
\name{export_DEG}
\alias{export_DEG}
\title{Export DEG results to CSV files}
\usage{
export_DEG(deg_list, outdir)

export_DEG(deg_list, outdir)
}
\arguments{
\item{deg_list}{Named list of DEG data.frames.}

\item{outdir}{Character output directory path.}
}
\description{
Export DEG results to CSV files

Export DEG to CSV
}
