% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core Functions.R
\name{generate_counts}
\alias{generate_counts}
\title{Generate count data with negative binomial model}
\usage{
generate_counts(meta, genes, base_means, dispersions, size_factors)

generate_counts(meta, genes, base_means, dispersions, size_factors)
}
\arguments{
\item{meta}{Data.frame of sample metadata.}

\item{genes}{Character vector of gene IDs.}

\item{base_means}{Numeric named vector of base expression means per gene.}

\item{dispersions}{Numeric named vector of dispersions per gene.}

\item{size_factors}{Numeric named vector of library size factors per sample.}
}
\value{
Tibble of counts for each gene-sample.
}
\description{
Generate count data with negative binomial model

Generate count matrix
}
