% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core Functions.R, R/visualization.R
\name{plot_PCA}
\alias{plot_PCA}
\title{PCA plot for DESeq2 dataset}
\usage{
plot_PCA(dds, intgroup = c("group_time", "batch"))

plot_PCA(dds, intgroup = c("group_time", "batch"))

plot_PCA(dds, intgroup = c("group_time", "batch"))
}
\arguments{
\item{dds}{A DESeqDataSet object.}

\item{intgroup}{Character vector of metadata columns for grouping.}
}
\value{
ggplot object.

A ggplot object.
}
\description{
PCA plot for DESeq2 dataset

PCA plot

PCA plot for a DESeq2 dataset
}
