% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core Functions.R, R/visualization.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot for DEG results}
\usage{
plot_volcano(df, title)

plot_volcano(df, title)

plot_volcano(df, title)
}
\arguments{
\item{df}{A data.frame with log2FoldChange, significance, expression columns.}

\item{title}{Character plot title.}
}
\value{
ggplot object.

A ggplot object.
}
\description{
Volcano plot for DEG results

Volcano plot

Volcano plot for DEG results
}
