% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core Functions.R, R/simulation.R
\name{simulate_samples}
\alias{simulate_samples}
\title{Simulate sample metadata for two organisms}
\usage{
simulate_samples(
  org,
  times = c("0h", "2h", "4h", "8h"),
  groups = if (org == "EC") c("less", "equal", "more", "all") else c("less", "equal",
    "more", "none"),
  batches = paste0("B", 1:2),
  reps = 4
)

simulate_samples(
  org,
  times = c("0h", "2h", "4h", "8h"),
  groups = if (org == "EC") c("less", "equal", "more", "all") else c("less", "equal",
    "more", "none"),
  batches = paste0("B", 1:2),
  reps = 4
)

simulate_samples(
  org,
  times = c("0h", "2h", "4h", "8h"),
  groups = if (org == "EC") c("less", "equal", "more", "all") else c("less", "equal",
    "more", "none"),
  batches = paste0("B", 1:2),
  reps = 4
)
}
\arguments{
\item{org}{Character: organism label ("EC" or "PP").}

\item{times}{Character vector of time points.}

\item{groups}{Character vector of group labels.}

\item{batches}{Character vector of batch labels.}

\item{reps}{Integer: number of replicates per condition.}
}
\value{
A data.frame of sample metadata.

A data.frame of sample metadata.
}
\description{
Simulate sample metadata for two organisms

Simulate sample metadata

Simulate sample metadata for two organisms
}
